api/requirements.txt
# === Базовые зависимости (общие с ботом) ===
openai==1.54.3
httpx==0.27.0
python-dotenv>=1.0.0
Pillow>=10.0.0
asyncpg>=0.29.0
aiohttp>=3.8.0
APScheduler>=3.10.4
pytz>=2023.3

# === API-специфичные зависимости ===
fastapi>=0.115.0
uvicorn[standard]>=0.30.0
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.4
bcrypt==4.0.1
python-multipart>=0.0.9
pydantic[email]>=2.0.0

# === Хранение фото ===
cloudinary>=1.36.0
